Clarified reuse of calculated regridder/weights
The regridder calculated by xESMF can already be input, but it was not clear. I added language to the docstrings to clarify that the regridder is the same as the weights. I also added a test to tests/test_em.py > test_grid_point_isel_Z() to ensure the regridder is being used when available by asserting a time speedup of 8x when the interpolation is performed a second time.