Use package (not interpolating) without xESMF installed
xESMF is not available on Windows currently. If it is not available, the user cannot run horizontal interpolation. Horizontal interpolation is now in a separate function called `interp_multi_dim()` and warnings and errors are now in place to catch when xESMF is not available.

There are also two tests now to check for running `select` when xESMF is not available but not needed, and when not available but needed.

Code authored by @yosoyjay in branch https://github.com/axiom-data-science/extract_model/compare/main...add-pyinterp-shim. I am bringing it in piecemeal.