Updates filter method to optionally preserve ROMS coordinates
If a ROMS dataset is filtered by standard_names, there is a chance that
many of the coordinate and auxiliary coordinate variables could be
omitted from the filtered dataset. Preserving the coordinate structure
is often desired, or required in order to produce derivative products
from the dataset.

This commit updates the filter method to optionally preserve the ROMS
horizontal coordinate structure in the filtered dataset, the commit also
adds an option for preserving the mask variables. Mask variables aren't
coordinate variables by CF conventions but they are critical for
accurately working with curvilinear grids that have undesirable features
like overlapping (in reference to a lon/lat coordinate reference
system), singularities, non-differentiable boundaries, etc.

This commit also improves the unit testing for the associated method
with specific support for testing ROMS datasets based explicitly on NOAA
Operational Forecast models.