Adds xarray backend to support loading unstructured grids
This commit adds a new xarray backend as a plugin to support loading
FVCOM and other unstructured gridded data from DAP and netCDF sources.
The backend enables clients to pass a mapping of variables to rename,
the variables are renamed after the xarray objects are loaded from the
store but before assembled into a dataset which aleviates issues arising
from unstructured grids that have coordinate variables with multiple
dimensions.