Ignore code coverage for JIT compiled function
The index_of_sorted function in the triangular_mesh module is a numba
just-in-time (JIT) compiled function. Numba reads in the function's
source code at runtime and compiles it into machine code or some other
intermediate but really fast interpreted code, so therefore the body of
that function is never actually run and would never be available for
code coverage.