Makes numba optional dependency by implementing index_of_sorted in numpy
Questions about the usefulness of including numba for such a narrow case
were brought up. This commit implements a solution for index_of_sorted
using pure numpy. The algorithmic complexity of this approach is
O(n lg n), whereas the numba implementation is O(lg n), but in practical
benchmarking the numpy implementation does much better for certain
cases, and numba does better when the haystack array is more random.

This commit also includes benchmarking in the tests.