Fix for test_grid_point_isel_Z non-determinism
This commit changes an assertion into a warning. The test initially made
an assertion about the timing taken, but on CI testing environments this
timing can't necessarily be reliable due to resource sharing and other
modern contention issues with CIs, so instead we simply issue a warning
that it was slower than expected.