Adds option to preload data for unstructured subsetting
When subsetting unstructured grids, a mask is ultimately needed for
processing, but internally to xarray if a mask is used for a selector,
it causes any fetching to fetch the entire variable even if a smaller
portion would do. This commit enables the client to pass preload=True
which will perform a more intelligent selection of the indices (isel),
then call ds.load() which will fetch the minimum selection needed for
subsetting. This radically improves performance.